ADD ELISP task.

+++ Rakefile
@@ -1,8 +1,15 @@
-$LOAD_PATH << File.dirname(File.expand_path(__FILE__)) + '/test'
-require 'test_helper'
+%w/test lib bin/.each do |lib|
+  $LOAD_PATH << File.dirname(File.expand_path(__FILE__)) + "/#{lib}"
+end
 require 'rake/testtask'
 require 'open-uri'
 require 'fileutils'
+require 'migemo'
+require 'migemo-convert'
+require 'migemo-index'
+require 'migemo-chars'
+require 'migemo-cache'
+
 ROOT = File.dirname(File.expand_path(__FILE__))
 
 SKK_DICT     = "data/SKK-JISYO.L"
@@ -11,24 +18,24 @@ MIGEMO_INDEX = 'data/migemo-dict.idx'
 MIGEMO_CHARS = 'data/migemo-chars'
 MIGEMO_CACHE = 'data/migemo-dict.cache'
 MIGEMO_INDEX_CACHE = 'data/migemo-dict.cache.idx'
+ELISP        = 'elisp/migemo.el'
 
-RESOURCES = [MIGEMO_DICT, MIGEMO_INDEX, MIGEMO_CHARS, MIGEMO_CACHE, MIGEMO_INDEX_CACHE]
+RESOURCES = [MIGEMO_DICT, MIGEMO_INDEX, MIGEMO_CHARS, MIGEMO_CACHE, MIGEMO_INDEX_CACHE, ELISP]
 
 task :default => :test
 task :setup   => RESOURCES
 task :clean do
   RESOURCES.each do |f|
-    if File.exists?(f)
-      p f
-      File.unlink(f)
-    end
+    File.unlink(f) if File.exists?(f)
   end
 end
 
 task :test => RESOURCES do
+  require 'test_helper'
   Rake::TestTask.new do |t|
     t.libs << "test"
     t.test_files = FileList['test/*.rb']
+    t.warning = true
   end
 end
 
@@ -60,4 +67,15 @@ end
 
 file MIGEMO_CACHE => [MIGEMO_CHARS, MIGEMO_DICT] do |t|
   Migemo::Cache.new(MIGEMO_DICT, MIGEMO_CHARS).generate.save(MIGEMO_CACHE)
+end
+
+file ELISP do |t|
+  lisp_in = nil
+  File.open(ELISP + '.in') do |f|
+    lisp_in = f.read
+  end
+  lisp_in.gsub!(/@pkgdatadir@/, ROOT + '/data')
+  File.open(ELISP, 'w') do |f|
+    f.write(lisp_in)
+  end
 end
